added Idbrahim angular project

+++ src/app/posts/posts.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { PostService } from '../post.service';
 import { Post } from '../post';
+import Swal from 'sweetalert2';
+
 
 @Component({
   selector: 'app-posts',
@@ -8,6 +10,8 @@ import { Post } from '../post';
   styleUrls: ['./posts.component.scss']
 })
 export class PostsComponent implements OnInit {
+
+  display: boolean = false;
 
   myPost: Post = {
     title: '',
@@ -32,7 +36,7 @@ export class PostsComponent implements OnInit {
     this.postService.persistPost(this.myPost).subscribe((res: Post) => {
 
 
-      this.posts = [this.myPost, ...this.posts]
+      this.posts = [res, ...this.posts]
       this.myPost = {
         title: '',
         body: ''
@@ -40,11 +44,41 @@ export class PostsComponent implements OnInit {
 
     })
 
+
   }
   destroyPost(id: number) {
-    this.postService.deletePost(id).subscribe(() => {
-      this.posts = this.posts.filter((post) => post.id !== id)
+
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'You will not be able to recover this imaginary file!',
+      icon: 'question',
+      showCancelButton: true,
+      confirmButtonText: 'Yes, delete it!',
+      cancelButtonText: 'No, keep it'
+    }).then((result) => {
+      if (result.value) {
+
+        this.postService.deletePost(id).subscribe(() => {
+          this.posts = this.posts.filter((post) => post.id !== id)
+
+          Swal.fire({
+            title: 'Deleted!',
+            text: 'Your imaginary file has been deleted.',
+            icon: 'success',
+            timer: 5000
+          })
+
+        })
+
+      }
     })
+  }
+  toggleForm() {
+    this.display = !this.display;
+  }
+  info(ngModel) {
+    console.log(ngModel);
+
   }
 
 }
